requirements.txt
langgraph>=0.2
langchain-anthropic>=0.3
langchain-core>=0.3
playwright>=1.44
google-api-python-client>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.2
httpx>=0.27
beautifulsoup4>=4.12
pyyaml>=6.0
python-dotenv>=1.0
